Fix something and add more feauture

+++ src/Function/AudioFunction/NowPlaying.js
@@ -1,8 +1,22 @@
 import MsgState from "../../States/AudioStates/MsgState.js";
+import VConnectionState from "../../States/VoiceStates/VConnectionState.js";
+import VoiceState from "../../States/VoiceStates/VoiceState.js";
 import Queues from "../../utils/ForAudio/Queue.js";
 import { musicEmbed } from "../../utils/textFormatter.js";
 
 const NowPlaying = async (msg) => {
+    if (!VoiceState.getVoiceChannel) {
+        msg.reply('You are not in Voice Channel');
+        return;
+    }
+    if (!VConnectionState.getVConnection) {
+        msg.reply('Bot not in Voice Channel');
+        return;
+    }
+    if (VoiceState.getVoiceChannel.id != VoiceState.getClientVoiceId) {
+        msg.reply('You are not in Same Voice Channel');
+        return;
+    }
     if (MsgState.getPrevPlayMsg) {
         MsgState.getPrevPlayMsg.delete();
     }
